feat(docs): Merge Docs into Lp
