--- apps/docs/.env.example
@@ -1,14 +1,0 @@
-# Generated by Shelve CLI
-NUXT_PRIVATE_ADMIN_EMAILS=your_value
-NUXT_PRIVATE_RESEND_API_KEY=your_value
-TURBO_TOKEN=your_value
-NUXT_SESSION_PASSWORD=your_value
-NUXT_OAUTH_GITHUB_CLIENT_SECRET=your_value
-NUXT_PRIVATE_ENCRYPTION_KEY=your_value
-TURBO_TEAM=your_value
-NUXT_OAUTH_GITHUB_CLIENT_ID=your_value
-NUXT_OAUTH_GOOGLE_CLIENT_ID=your_value
-NUXT_PRIVATE_REDIS_URL=your_value
-DATABASE_URL=your_value
-NUXT_OAUTH_GOOGLE_CLIENT_SECRET=your_value
-NUXT_UI_PRO_LICENSE=your_value
