--- apps/docs/app/layouts/default.vue
@@ -1,26 +1,0 @@
-<script setup lang="ts">
-import type { ContentNavigationItem } from '@nuxt/content'
-
-const navigation = inject<Ref<ContentNavigationItem[]>>('navigation')
-</script>
-
-<template>
-  <UMain>
-    <UContainer>
-      <UPage>
-        <template #left>
-          <UPageAside>
-            <UContentNavigation :navigation highlight :ui="{ linkTrailingBadge: 'font-semibold uppercase' }">
-              <template #link-title="{ link }">
-                <span class="inline-flex items-center gap-0.5">
-                  {{ link.title }}
-                </span>
-              </template>
-            </UContentNavigation>
-          </UPageAside>
-        </template>
-        <slot />
-      </UPage>
-    </UContainer>
-  </UMain>
-</template>
