--- apps/docs/eslint.config.js
@@ -1,3 +1,0 @@
-import { createConfig } from "@hrcd/eslint-config"
-
-export default createConfig({})
