--- apps/docs/package.json
@@ -1,24 +1,0 @@
-{
-  "name": "@shelve/docs",
-  "private": true,
-  "type": "module",
-  "scripts": {
-    "build": "nuxt build",
-    "dev": "nuxt dev",
-    "generate": "nuxt generate",
-    "preview": "nuxt preview",
-    "start": "node .output/server/index.mjs",
-    "postinstall": "nuxt prepare",
-    "lint": "eslint .",
-    "lint:fix": "eslint --fix .",
-    "typecheck": "tsc --noEmit"
-  },
-  "dependencies": {
-    "@nuxt/content": "3.0.0-alpha.8",
-    "@nuxt/scripts": "0.9.5",
-    "@nuxt/ui-pro": "3.0.0-alpha.10",
-    "nuxt": "3.15.1",
-    "nuxt-og-image": "^4.0.2",
-    "vue": "3.5.13"
-  }
-}
