--- apps/docs/public/robots.txt
@@ -1,1 +1,0 @@
-user-agent: *
