--- apps/docs/server/api/hello.ts
@@ -1,3 +1,0 @@
-export default defineEventHandler(() => {
-  return 'Hello World!'
-})
