--- apps/docs/shelve.json
@@ -1,4 +1,0 @@
-{
-  "slug": "hrcd",
-  "project": "shelve"
-}
