+++ apps/lp/app/app.config.ts
@@ -1,0 +1,8 @@
+export default defineAppConfig({
+  ui: {
+    icons: {
+      light: 'heroicons:moon',
+      dark: 'heroicons:sun'
+    }
+  }
+})
