+++ apps/lp/app/assets/css/index.css
