+++ apps/lp/content.config.ts
