+++ apps/lp/content/1.getting-started/.navigation.yml
