+++ apps/lp/content/1.getting-started/1.index.md
