+++ apps/lp/content/2.cli/1.index.md
