+++ apps/lp/package.json
@@ -20,11 +20,12 @@
     "@iconify-json/simple-icons": "1.2.19",
     "@shelve/types": "*",
     "@nuxt/image": "1.9.0",
-    "@nuxt/scripts": "0.9.5",
-    "@nuxt/ui": "3.0.0-alpha.10",
     "@nuxtjs/seo": "2.0.2",
     "nuxt": "3.15.1",
     "nuxt-build-cache": "0.1.1",
+    "@nuxt/content": "3.0.0-alpha.8",
+    "@nuxt/scripts": "0.9.5",
+    "@nuxt/ui-pro": "3.0.0-alpha.10",
     "vue": "3.5.13",
     "vue-router": "4.5.0",
     "vue-sonner": "1.3.0"
