--- docker/docs-compose.coolify.yml
@@ -1,39 +1,0 @@
-# documentation: https://docs.shelve.cloud
-# slogan: Shelve - Secure environment sharing for collaborative development
-# tags: environment variables, projects management, collaboration, cli, secure sharing
-# port: 3000
-
-services:
-  shelve_docs:
-    image: ghcr.io/hugorcd/shelve-docs:latest
-    container_name: docs
-    restart: always
-    environment:
-      - NUXT_UI_PRO_LICENSE=${NUXT_UI_PRO_LICENSE}
-    ports:
-      - "3000:3000"
-    healthcheck:
-      test: [ "CMD", "curl", "-f", "http://localhost:3000/api/hello" ]
-      interval: 30s
-      timeout: 10s
-      retries: 3
-    logging:
-      driver: "json-file"
-      options:
-        max-size: "10m"
-        max-file: "3"
-    deploy:
-      resources:
-        limits:
-          cpus: '1'
-          memory: 1G
-        reservations:
-          cpus: '0.5'
-          memory: 512M
-    networks:
-      - shelve-network
-
-networks:
-  shelve-network:
-    driver: bridge
-
