+++ package.json
@@ -7,12 +7,10 @@
     "build": "turbo build",
     "build:app": "turbo build --filter=@shelve/app",
     "build:lp": "turbo build --filter=@shelve/lp",
-    "build:docs": "turbo build --filter=@shelve/docs",
     "build:cli": "turbo build --filter=@shelve/cli",
     "dev": "turbo dev --filter='./apps/*'",
     "dev:app": "turbo dev --filter=@shelve/{app,types}",
     "dev:lp": "turbo dev --filter=@shelve/{lp,types}",
-    "dev:docs": "turbo dev --filter=@shelve/docs",
     "dev:cli": "turbo dev --filter=@shelve/cli",
     "release": "turbo release && changeset",
     "dev:prepare": "turbo dev:prepare",
